[BugFix] Fix incorrect time unit for JournalTask#betterCommitBeforeTime (#20372)
`JournalTask#betterCommitBeforeTime` is assigned a millisecond time, but used as a nano time in `JournalWriter#shouldCommitNow`. This bug has no bad effect currently because `JournalTask` is always constructed with a -1 `maxWaitIntervalMs`, and it causes `betterCommitBeforeTime` is also -1 which means an invalid time. This problem could be fixed by assigning a nano time to `betterCommitBeforeTime`

Signed-off-by: PengFei Li <lpengfei2016@gmail.com>