[Enhancement] Add iceberg partition prune info when executing explain verbose query (#21148)
Signed-off-by: stephen <stephen5217@163.com>
If we want to get all existed partitions in an iceberg table, fe need to access all manifest metadata files. This is time-consuming and it will take a few minutes for some corner cases. So we removed the partition information for explain iceberg.
But if user want to check if partition prune is effect, we can't find. So we support explain verbose to display partition information in order to help user optimize sql. At the same time, normal explain is guaranteed to return quickly.