[Feature] support key partition exchanger for external table sink (#22115)
Signed-off-by: stephen <stephen5217@163.com>

When writing data to an Iceberg partitioned table with an insert statement, values for the same partition column should be written to the same file. And an iceberg sink operator could include multiple partitioned writers. In order to reduce small files, the partitions processed by each pipeline driver are orthogonal. If I use the current PartitionExchanger, each chunk accepted by an iceberg_sink_operator may contain rows of multiple partitions. So I need to perform a second shuffle. This result memory copy and lower performance.

In summary, supporting lake partition exchange is mainly to ensure that the partition column values of each chunk which iceberg_sink_operator accepted are the same. and it only once shuffle when iceberg table sink writes to multiple partitions.