[Enhancement] Move the check from the description to the comment (#23647)
Because GitHub places the `PULL_REQUEST_TEMPLATE.md` in the description
of the pull request. When we squash the pull request, many unused checks
will be in the `git log`.
This pull request moves the checks from the description to the comment.
Then the unused checks will don't include in the `git log`.