[Enhancement] Trigger statistics collection after first load (#23694)
## Problem Summary:

Currently, full statistics collection will be performed for all tables
with an interval of 5min(defined by
`Config::statistic_collect_interval_sec`). But in some scenerios, this
timeliness is not enough. For example, a ETL process may contain
multiply intermediate tables, which will be accessed right after load,
and at this moment, statistics has not been generated, so it may produce
terrible execution plan.

In order to solve this problem, we're trying to trigger a full/sample
statistics collection right after the first loading process is finished.
* If loaded data row count >= `statistic_sample_collect_rows`, a sample
statistics collection will be triggered
* otherwise, a full statistics collection will be triggered.

There are 5 ways for loading data, which respectively are:
1. `insert into`
2. `stream load`
3. `broker load`
4. `spark load`
5. `routine load`

And the above 5 approaches can be categorized into 3 groups:
1. `insert into`
2. `stream load`
3. `broker load/spark load/routine load`

For `insert into`, we can add a sync process to trigger statistics
collection process and hold execution until the collection is finished.

For `stream load`, this kind of load approach is totally controled by a
coordinator be, and fe will finish this transaction aynscly. So the
statistics collection process must be async if we don't want to change
the whole stream load procedure.

For `broker load/spark load/routine load`, these three load approaches
are async and are constraint by load task management, and we can check
the loading status by the command `show load`. So we just need to make
sure that the progress of the load task won't become 100 until the
statistics collection is finished.

### Config

A configuration `enable_statistic_collect_on_first_load` has been added,
with a default value true, to enable/disable this feature.

## What type of PR is this:
- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

## Checklist:
- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr will affect users' behaviors
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function

## Bugfix cherry-pick branch check:
- [x] I have checked the version labels which the pr will be auto
backported to target branch
  - [ ] 3.0
  - [ ] 2.5
  - [ ] 2.4
  - [ ] 2.3

---------

Signed-off-by: liuyehcf <1559500551@qq.com>